requirements.txt
fastapi
uvicorn
python-dotenv
aiofiles
asyncpg
aiosqlite
transformers
torch
openai
pydantic
passlib[bcrypt]
pyjwt
